refactor: unify OpenRouter HTTP client in SpeakCore
Move the OpenRouter API client from SpeakApp into SpeakCore as a single
public, platform-neutral client (Foundation/URLSession + AVFoundation for
audio duration only) and migrate all macOS and iOS call sites to it:

- SpeakCore/OpenRouterAPIClient: public actor with lazy API-key resolution
  via an async provider closure, streaming SSE chat, non-streaming chat
  (now with optional max_tokens), inline-audio batch transcription,
  key validation (debug snapshots stay redacted via
  APIKeyValidationDebugSnapshot/SensitiveHeaderRedactor), local fallbacks,
  and connection warm-up. Streaming now uses the injected URLSession and
  cancels the network task when the stream is terminated.
- SpeakApp: delete the old client; add a thin SecureAppStorage-backed
  convenience init so WireUp/TranscriptionManager/PostProcessingManager/
  MainManager keep compiling unchanged.
- SpeakiOS: delete the inline SSE re-implementation in
  iOSPostProcessingManager, the raw openrouter.ai call in VoiceSummariser,
  and the duplicated transcription payload builder in iOSBatchTranscriber;
  all now go through the shared client.
- Tests: move client tests to SpeakCoreTests (no keychain dependency),
  add SSE stream-parsing coverage (ordered chunks, [DONE], keep-alive
  comments, role-only deltas, missing key), and keep the provider-registry
  routing tests in SpeakAppTests.

Behaviour notes: iOS OpenRouter requests now send the unified
platform branding headers; VoiceSummariser keeps its exact headers and
error mapping. iOS-only views/services still need iOS CI to confirm.
