--- Sources/SpeakApp/OpenRouterAPIClient.swift
@@ -1,874 +1,0 @@
-import SpeakCore
-import AVFoundation
-import Foundation
-import os.log
-
-enum OpenRouterClientError: LocalizedError {
-  case apiKeyMissing
-  case invalidResponse
-  case httpStatus(Int, String)
-  case audioFileTooLarge(fileSize: Int64, limit: Int64)
-
-  var errorDescription: String? {
-    switch self {
-    case .apiKeyMissing:
-      return "OpenRouter API key is missing."
-    case .invalidResponse:
-      return "The server returned an invalid response."
-    case .httpStatus(let code, let body):
-      return "OpenRouter responded with status \(code): \(body)"
-    case .audioFileTooLarge(let fileSize, let limit):
-      let fileSizeDescription = ByteCountFormatter.string(fromByteCount: fileSize, countStyle: .file)
-      let limitDescription = ByteCountFormatter.string(fromByteCount: limit, countStyle: .file)
-      return "Audio file is too large for OpenRouter reprocessing (\(fileSizeDescription), limit \(limitDescription))."
-    }
-  }
-}
-
-actor OpenRouterAPIClient: StreamingChatLLMClient, BatchTranscriptionClient { // swiftlint:disable:this type_body_length
-  private let baseURL = URL(string: "https://openrouter.ai/api/v1")!
-  private let session: URLSession
-  private let secureStorage: SecureAppStorage
-  private let apiKeyOverride: String?
-  private let apiKeyIdentifier: String
-  private let maximumInlineAudioBytes: Int64
-  private let logger = Logger(subsystem: "com.github.speakapp", category: "OpenRouter")
-  private let titleHeaderValue = "SpeakApp (macOS)"
-  private let refererHeaderValue = "https://github.com/speak"
-
-  private struct ValidationAttemptResult {
-    let success: Bool
-    let debug: APIKeyValidationDebugSnapshot
-    let message: String
-  }
-
-  private struct ValidationAttemptError: Error {
-    let debug: APIKeyValidationDebugSnapshot
-    let message: String
-  }
-
-  init(
-    secureStorage: SecureAppStorage,
-    session: URLSession = .shared,
-    apiKeyOverride: String? = nil,
-    apiKeyIdentifier: String = "openrouter.apiKey",
-    maximumInlineAudioBytes: Int64 = 50 * 1024 * 1024
-  ) {
-    self.secureStorage = secureStorage
-    self.session = session
-    self.apiKeyOverride = apiKeyOverride
-    self.apiKeyIdentifier = apiKeyIdentifier
-    self.maximumInlineAudioBytes = maximumInlineAudioBytes
-  }
-
-  func hasStoredAPIKey() async -> Bool {
-    await storedAPIKey() != nil
-  }
-
-  func requiresRemoteAccess(for model: String) -> Bool {
-    !allowsLocalFallback(for: model)
-  }
-
-  func sendChat(
-    systemPrompt: String?,
-    messages: [ChatMessage],
-    model: String,
-    temperature: Double
-  ) async throws -> ChatResponse {
-    if let key = await storedAPIKey() {
-      return try await performRemoteChat(
-        apiKey: key,
-        systemPrompt: systemPrompt,
-        messages: messages,
-        model: model,
-        temperature: temperature
-      )
-    }
-
-    return performLocalChatFallback(systemPrompt: systemPrompt, messages: messages)
-  }
-
-  nonisolated func sendChatStreaming( // swiftlint:disable:this cyclomatic_complexity function_body_length
-    systemPrompt: String?,
-    messages: [ChatMessage],
-    model: String,
-    temperature: Double
-  ) -> AsyncThrowingStream<String, Error> {
-    AsyncThrowingStream { continuation in
-      Task {
-        do {
-          guard let key = await self.storedAPIKey() else {
-            continuation.finish(throwing: OpenRouterClientError.apiKeyMissing)
-            return
-          }
-
-          let url = self.baseURL.appendingPathComponent("chat/completions")
-          var request = URLRequest(url: url)
-          request.httpMethod = "POST"
-          request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-          request.setValue("Bearer \(key)", forHTTPHeaderField: "Authorization")
-          await self.applyBrandHeaders(&request)
-
-          let builtMessages = await self.buildMessages(systemPrompt: systemPrompt, messages: messages)
-          let streamingMessages = builtMessages.map { msg in
-            OpenRouterStreamingChatRequest.Message(role: msg.role, content: msg.content)
-          }
-          let payload = OpenRouterStreamingChatRequest(
-            model: model,
-            temperature: temperature,
-            messages: streamingMessages,
-            stream: true
-          )
-
-          request.httpBody = try JSONEncoder().encode(payload)
-
-          let (bytes, response) = try await URLSession.shared.bytes(for: request)
-          guard let http = response as? HTTPURLResponse else {
-            continuation.finish(throwing: OpenRouterClientError.invalidResponse)
-            return
-          }
-          guard (200..<300).contains(http.statusCode) else {
-            var body = ""
-            for try await line in bytes.lines {
-              body += line
-            }
-            continuation.finish(throwing: OpenRouterClientError.httpStatus(http.statusCode, body))
-            return
-          }
-
-          for try await line in bytes.lines {
-            guard !line.isEmpty else { continue }
-            guard line.hasPrefix("data: ") else { continue }
-
-            let jsonString = String(line.dropFirst(6))
-            if jsonString == "[DONE]" { break }
-
-            guard let data = jsonString.data(using: .utf8) else { continue }
-            if let chunk = try? JSONDecoder().decode(OpenRouterStreamChunk.self, from: data),
-               let delta = chunk.choices.first?.delta,
-               let content = delta.content, !content.isEmpty {
-              continuation.yield(content)
-            }
-          }
-          continuation.finish()
-        } catch {
-          continuation.finish(throwing: error)
-        }
-      }
-    }
-  }
-
-  func transcribeFile(at url: URL, model: String, language: String?) async throws
-    -> TranscriptionResult
-  {
-    let cleanedModel = model.trimmingCharacters(in: .whitespacesAndNewlines)
-
-    if let key = await storedAPIKey() {
-      return try await performRemoteTranscription(
-        apiKey: key, url: url, model: cleanedModel, language: language)
-    }
-
-    if allowsLocalFallback(for: cleanedModel) {
-      return try await localTranscriptionFallback(url: url, model: cleanedModel)
-    }
-
-    throw OpenRouterClientError.apiKeyMissing
-  }
-
-  private func storedAPIKey() async -> String? {
-    if let apiKeyOverride {
-      let trimmed = apiKeyOverride.trimmingCharacters(in: .whitespacesAndNewlines)
-      if !trimmed.isEmpty {
-        return trimmed
-      }
-    }
-
-    let rawKey = try? await secureStorage.secret(identifier: apiKeyIdentifier)
-    let key = rawKey?.trimmingCharacters(in: .whitespacesAndNewlines)
-    return key.flatMap { $0.isEmpty ? nil : $0 }
-  }
-
-  func validateAPIKey(_ key: String) async -> APIKeyValidationResult {
-    let trimmed = key.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else {
-      return .failure(
-        message: "API key is empty",
-        debug: APIKeyValidationDebugSnapshot(
-          url: "",
-          method: "",
-          requestHeaders: [:],
-          requestBody: nil,
-          statusCode: nil,
-          responseHeaders: [:],
-          responseBody: nil,
-          errorDescription: "API key is empty"
-        )
-      )
-    }
-
-    var lastFailure: (message: String, debug: APIKeyValidationDebugSnapshot?) =
-      ("Validation failed", nil)
-
-    do {
-      let authResult = try await validateViaAuthEndpoint(apiKey: trimmed)
-      if authResult.success {
-        return .success(
-          message: "OpenRouter API key validated via auth endpoint",
-          debug: authResult.debug
-        )
-      }
-      lastFailure = (authResult.message, authResult.debug)
-    } catch let attemptError as ValidationAttemptError {
-      logger.error("API key auth validation failed: \(attemptError.message, privacy: .public)")
-      lastFailure = (attemptError.message, attemptError.debug)
-    } catch {
-      logger.error("API key auth validation failed: \(error.localizedDescription, privacy: .public)")
-      let debug = APIKeyValidationDebugSnapshot(
-        url: baseURL.appendingPathComponent("auth/validate").absoluteString,
-        method: "GET",
-        requestHeaders: [:],
-        requestBody: nil,
-        statusCode: nil,
-        responseHeaders: [:],
-        responseBody: nil,
-        errorDescription: error.localizedDescription
-      )
-      lastFailure = (error.localizedDescription, debug)
-    }
-
-    do {
-      let chatResult = try await validateViaChat(apiKey: trimmed)
-      if chatResult.success {
-        return .success(
-          message: "OpenRouter API key validated via models endpoint",
-          debug: chatResult.debug
-        )
-      }
-      lastFailure = (chatResult.message, chatResult.debug)
-    } catch let attemptError as ValidationAttemptError {
-      logger.error(
-        "API key validation via chat failed: \(attemptError.message, privacy: .public)"
-      )
-      lastFailure = (attemptError.message, attemptError.debug)
-    } catch {
-      logger.error(
-        "API key validation via chat failed: \(error.localizedDescription, privacy: .public)"
-      )
-      let debug = APIKeyValidationDebugSnapshot(
-        url: baseURL.appendingPathComponent("models").absoluteString,
-        method: "GET",
-        requestHeaders: [:],
-        requestBody: nil,
-        statusCode: nil,
-        responseHeaders: [:],
-        responseBody: nil,
-        errorDescription: error.localizedDescription
-      )
-      lastFailure = (error.localizedDescription, debug)
-    }
-
-    return .failure(message: lastFailure.message, debug: lastFailure.debug)
-  }
-
-  private func validateViaAuthEndpoint(apiKey: String) async throws -> ValidationAttemptResult {
-    let url = baseURL.appendingPathComponent("auth/validate")
-    var request = URLRequest(url: url)
-    request.httpMethod = "GET"
-    request.setValue("Bearer \(apiKey)", forHTTPHeaderField: "Authorization")
-    applyBrandHeaders(&request)
-
-    let headers = request.allHTTPHeaderFields ?? [:]
-
-    do {
-      let (data, response) = try await session.data(for: request)
-      guard let http = response as? HTTPURLResponse else {
-        let debug = APIKeyValidationDebugSnapshot(
-          url: url.absoluteString,
-          method: request.httpMethod ?? "GET",
-          requestHeaders: headers,
-          requestBody: nil,
-          statusCode: nil,
-          responseHeaders: [:],
-          responseBody: nil,
-          errorDescription: OpenRouterClientError.invalidResponse.localizedDescription
-        )
-        throw ValidationAttemptError(
-          debug: debug,
-          message: OpenRouterClientError.invalidResponse.localizedDescription
-        )
-      }
-
-      let responseBody = string(from: data)
-      let debug = APIKeyValidationDebugSnapshot(
-        url: url.absoluteString,
-        method: request.httpMethod ?? "GET",
-        requestHeaders: headers,
-        requestBody: nil,
-        statusCode: http.statusCode,
-        responseHeaders: normalizedHeaders(http.allHeaderFields),
-        responseBody: responseBody,
-        errorDescription: http.statusCode == 200 ? nil : responseBody
-      )
-
-      if http.statusCode == 200 {
-        if let payload = try? JSONDecoder().decode(OpenRouterValidationResponse.self, from: data) {
-          if let valid = payload.valid {
-            return ValidationAttemptResult(
-              success: valid,
-              debug: debug,
-              message: valid ? "Validated" : "Auth validation returned invalid"
-            )
-          }
-          if let valid = payload.data?.valid {
-            return ValidationAttemptResult(
-              success: valid,
-              debug: debug,
-              message: valid ? "Validated" : "Auth validation returned invalid"
-            )
-          }
-        }
-        return ValidationAttemptResult(
-          success: true,
-          debug: debug,
-          message: "Validated"
-        )
-      }
-
-      if http.statusCode == 401 {
-        return ValidationAttemptResult(
-          success: false,
-          debug: debug,
-          message: "Unauthorized response from auth endpoint"
-        )
-      }
-
-      throw ValidationAttemptError(
-        debug: debug,
-        message: "Unexpected status \(http.statusCode)"
-      )
-    } catch let attemptError as ValidationAttemptError {
-      throw attemptError
-    } catch {
-      let debug = APIKeyValidationDebugSnapshot(
-        url: url.absoluteString,
-        method: request.httpMethod ?? "GET",
-        requestHeaders: headers,
-        requestBody: nil,
-        statusCode: nil,
-        responseHeaders: [:],
-        responseBody: nil,
-        errorDescription: error.localizedDescription
-      )
-      throw ValidationAttemptError(debug: debug, message: error.localizedDescription)
-    }
-  }
-
-  private func validateViaChat(apiKey: String) async throws -> ValidationAttemptResult {
-    let url = baseURL.appendingPathComponent("models")
-    var request = URLRequest(url: url)
-    request.httpMethod = "GET"
-    request.setValue("Bearer \(apiKey)", forHTTPHeaderField: "Authorization")
-    applyBrandHeaders(&request)
-
-    let headers = request.allHTTPHeaderFields ?? [:]
-
-    do {
-      let (data, response) = try await session.data(for: request)
-      guard let http = response as? HTTPURLResponse else {
-        let debug = APIKeyValidationDebugSnapshot(
-          url: url.absoluteString,
-          method: request.httpMethod ?? "GET",
-          requestHeaders: headers,
-          requestBody: nil,
-          statusCode: nil,
-          responseHeaders: [:],
-          responseBody: nil,
-          errorDescription: OpenRouterClientError.invalidResponse.localizedDescription
-        )
-        throw ValidationAttemptError(
-          debug: debug,
-          message: OpenRouterClientError.invalidResponse.localizedDescription
-        )
-      }
-
-      let responseBody = string(from: data)
-      let debug = APIKeyValidationDebugSnapshot(
-        url: url.absoluteString,
-        method: request.httpMethod ?? "GET",
-        requestHeaders: headers,
-        requestBody: nil,
-        statusCode: http.statusCode,
-        responseHeaders: normalizedHeaders(http.allHeaderFields),
-        responseBody: responseBody,
-        errorDescription: (200..<300).contains(http.statusCode) ? nil : responseBody
-      )
-
-      if (200..<300).contains(http.statusCode) {
-        return ValidationAttemptResult(
-          success: true,
-          debug: debug,
-          message: "Validated"
-        )
-      }
-
-      if http.statusCode == 401 {
-        return ValidationAttemptResult(
-          success: false,
-          debug: debug,
-          message: "Unauthorized response from models endpoint"
-        )
-      }
-
-      throw ValidationAttemptError(
-        debug: debug,
-        message: "Unexpected status \(http.statusCode)"
-      )
-    } catch let attemptError as ValidationAttemptError {
-      throw attemptError
-    } catch {
-      let debug = APIKeyValidationDebugSnapshot(
-        url: url.absoluteString,
-        method: request.httpMethod ?? "GET",
-        requestHeaders: headers,
-        requestBody: nil,
-        statusCode: nil,
-        responseHeaders: [:],
-        responseBody: nil,
-        errorDescription: error.localizedDescription
-      )
-      throw ValidationAttemptError(debug: debug, message: error.localizedDescription)
-    }
-  }
-
-  private func performRemoteChat(
-    apiKey: String,
-    systemPrompt: String?,
-    messages: [ChatMessage],
-    model: String,
-    temperature: Double
-  ) async throws -> ChatResponse {
-    let url = baseURL.appendingPathComponent("chat/completions")
-    var request = URLRequest(url: url)
-    request.httpMethod = "POST"
-    request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-    request.setValue("Bearer \(apiKey)", forHTTPHeaderField: "Authorization")
-    applyBrandHeaders(&request)
-
-    let payload = OpenRouterChatRequest(
-      model: model,
-      temperature: temperature,
-      messages: buildMessages(systemPrompt: systemPrompt, messages: messages)
-    )
-
-    request.httpBody = try JSONEncoder().encode(payload)
-
-    let (data, response) = try await session.data(for: request)
-    guard let http = response as? HTTPURLResponse else {
-      throw OpenRouterClientError.invalidResponse
-    }
-    guard (200..<300).contains(http.statusCode) else {
-      let body = String(data: data, encoding: .utf8) ?? "<no-body>"
-      throw OpenRouterClientError.httpStatus(http.statusCode, body)
-    }
-
-    let decoded = try JSONDecoder().decode(OpenRouterChatResponse.self, from: data)
-    let assistantMessages = decoded.choices.compactMap { choice in
-      choice.message.map { ChatMessage(role: .assistant, content: $0.content) }
-    }
-
-    let finishReason = decoded.choices.first?.finish_reason ?? "stop"
-    let cost = decoded.usage.map { usage in
-      ChatCostBreakdown(
-        inputTokens: usage.prompt_tokens,
-        outputTokens: usage.completion_tokens,
-        totalCost: Decimal(usage.prompt_tokens + usage.completion_tokens) / 1_000_000,
-        currency: "USD"
-      )
-    }
-
-    var conversation: [ChatMessage] = []
-    if let systemPrompt {
-      conversation.append(ChatMessage(role: .system, content: systemPrompt))
-    }
-    conversation.append(contentsOf: messages)
-    conversation.append(contentsOf: assistantMessages)
-
-    return ChatResponse(
-      messages: conversation,
-      finishReason: finishReason,
-      cost: cost,
-      rawPayload: String(data: data, encoding: .utf8)
-    )
-  }
-
-  private func performRemoteTranscription(
-    apiKey: String,
-    url: URL,
-    model: String,
-    language: String?
-  ) async throws -> TranscriptionResult {
-    let endpoint = baseURL.appendingPathComponent("chat/completions")
-    var request = URLRequest(url: endpoint)
-    request.httpMethod = "POST"
-    request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-    request.setValue("Bearer \(apiKey)", forHTTPHeaderField: "Authorization")
-    applyBrandHeaders(&request)
-
-    try enforceInlineAudioSizeLimit(for: url)
-    let audioData = try Data(contentsOf: url)
-    let prompt = transcriptionPrompt(language: language)
-    let payload = OpenRouterAudioTranscriptionRequest(
-      model: model,
-      temperature: 0,
-      messages: [
-        OpenRouterAudioTranscriptionRequest.Message(
-          role: "user",
-          content: [
-            .text(prompt),
-            .inputAudio(
-              data: audioData.base64EncodedString(),
-              format: audioInputFormat(for: url)
-            )
-          ]
-        )
-      ],
-      stream: false
-    )
-    request.httpBody = try JSONEncoder().encode(payload)
-
-    let (data, response) = try await session.data(for: request)
-    guard let http = response as? HTTPURLResponse else {
-      throw OpenRouterClientError.invalidResponse
-    }
-    guard (200..<300).contains(http.statusCode) else {
-      let body = String(data: data, encoding: .utf8) ?? "<no-body>"
-      throw OpenRouterClientError.httpStatus(http.statusCode, body)
-    }
-
-    let decoded = try JSONDecoder().decode(OpenRouterChatResponse.self, from: data)
-    guard
-      let text = decoded.choices
-        .compactMap({ $0.message?.content.trimmingCharacters(in: .whitespacesAndNewlines) })
-        .first(where: { !$0.isEmpty })
-    else {
-      throw OpenRouterClientError.invalidResponse
-    }
-
-    return try await buildTranscriptionResult(
-      text: text,
-      audioURL: url,
-      model: model,
-      payload: data
-    )
-  }
-
-  private func localTranscriptionFallback(url: URL, model: String) async throws
-    -> TranscriptionResult
-  {
-    let asset = AVURLAsset(url: url)
-    let durationTime = try await asset.load(.duration)
-    let duration = durationTime.seconds
-    let text = "Transcription placeholder for \(url.lastPathComponent)"
-    let segment = TranscriptionSegment(startTime: 0, endTime: duration, text: text)
-    return TranscriptionResult(
-      text: text,
-      segments: [segment],
-      confidence: nil,
-      duration: duration,
-      modelIdentifier: model,
-      cost: nil,
-      rawPayload: nil,
-      debugInfo: nil
-    )
-  }
-
-  private func performLocalChatFallback(systemPrompt: String?, messages: [ChatMessage])
-    -> ChatResponse
-  {
-    let userText =
-      messages.last(where: { $0.role == .user })?.content
-      ?? messages.last?.content
-      ?? ""
-    let processed = heuristicPostProcess(text: userText)
-
-    var conversation: [ChatMessage] = []
-    if let systemPrompt {
-      conversation.append(ChatMessage(role: .system, content: systemPrompt))
-    }
-    conversation.append(contentsOf: messages)
-    conversation.append(ChatMessage(role: .assistant, content: processed))
-
-    return ChatResponse(
-      messages: conversation, finishReason: "fallback-local", cost: nil, rawPayload: nil)
-  }
-
-  private func heuristicPostProcess(text: String) -> String {
-    let trimmed = text.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return text }
-
-    let collapsed =
-      trimmed
-      .components(separatedBy: .whitespacesAndNewlines)
-      .filter { !$0.isEmpty }
-      .joined(separator: " ")
-
-    guard !collapsed.isEmpty else { return trimmed }
-
-    var result = collapsed
-    if let first = result.first {
-      let uppercaseFirst = String(first).uppercased()
-      result.replaceSubrange(result.startIndex...result.startIndex, with: uppercaseFirst)
-    }
-
-    if let last = result.last, !".!?".contains(last) {
-      result.append(".")
-    }
-
-    return result
-  }
-
-  private func buildMessages(systemPrompt: String?, messages: [ChatMessage])
-    -> [OpenRouterChatRequest.Message]
-  {
-    var payload: [OpenRouterChatRequest.Message] = []
-    if let systemPrompt {
-      payload.append(.init(role: "system", content: systemPrompt))
-    }
-    payload += messages.map { message in
-      .init(role: message.role.rawValue, content: message.content)
-    }
-    return payload
-  }
-
-  private func string(from data: Data?) -> String? {
-    guard let data else { return nil }
-    if let text = String(data: data, encoding: .utf8) {
-      return text
-    }
-    return String(data: data, encoding: .ascii)
-  }
-
-  private func normalizedHeaders(_ headers: [AnyHashable: Any]) -> [String: String] {
-    var normalized: [String: String] = [:]
-    for (key, value) in headers {
-      guard let keyString = key as? String else { continue }
-      if let valueString = value as? String {
-        normalized[keyString] = valueString
-      } else {
-        normalized[keyString] = String(describing: value)
-      }
-    }
-    return normalized
-  }
-
-  private func transcriptionPrompt(language: String?) -> String {
-    let trimmedLanguage = language?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    if trimmedLanguage.isEmpty {
-      return "Transcribe this audio file. Return only the transcript text, with no commentary."
-    }
-
-    return "Transcribe this audio file using locale \(trimmedLanguage). "
-      + "Return only the transcript text, with no commentary."
-  }
-
-  private func audioInputFormat(for url: URL) -> String {
-    let ext = url.pathExtension.lowercased()
-    switch ext {
-    case "wav", "mp3", "aiff", "aac", "ogg", "flac", "m4a", "pcm16", "pcm24":
-      return ext
-    case "m4b":
-      return "m4a"
-    case "wave":
-      return "wav"
-    default:
-      return "m4a"
-    }
-  }
-
-  private func enforceInlineAudioSizeLimit(for url: URL) throws {
-    let fileSize = try audioFileSize(for: url)
-    guard fileSize <= maximumInlineAudioBytes else {
-      throw OpenRouterClientError.audioFileTooLarge(
-        fileSize: fileSize,
-        limit: maximumInlineAudioBytes
-      )
-    }
-  }
-
-  private func audioFileSize(for url: URL) throws -> Int64 {
-    if let fileSize = try url.resourceValues(forKeys: [.fileSizeKey]).fileSize {
-      return Int64(fileSize)
-    }
-
-    let attributes = try FileManager.default.attributesOfItem(atPath: url.path)
-    return (attributes[.size] as? NSNumber)?.int64Value ?? 0
-  }
-
-  private func buildTranscriptionResult(
-    text: String,
-    audioURL: URL,
-    model: String,
-    payload: Data
-  ) async throws -> TranscriptionResult {
-    let asset = AVURLAsset(url: audioURL)
-    let durationTime = try await asset.load(.duration)
-    let duration = durationTime.seconds
-    let segments = [TranscriptionSegment(startTime: 0, endTime: duration, text: text)]
-
-    return TranscriptionResult(
-      text: text,
-      segments: segments,
-      confidence: nil,
-      duration: duration,
-      modelIdentifier: model,
-      cost: nil,
-      rawPayload: String(data: payload, encoding: .utf8),
-      debugInfo: nil
-    )
-  }
-
-  private func applyBrandHeaders(_ request: inout URLRequest) {
-    request.setValue(titleHeaderValue, forHTTPHeaderField: "X-Title")
-    request.setValue(refererHeaderValue, forHTTPHeaderField: "HTTP-Referer")
-    request.setValue(refererHeaderValue, forHTTPHeaderField: "Referer")
-  }
-
-  private func allowsLocalFallback(for model: String) -> Bool {
-    if model.lowercased().hasPrefix("apple/") { return true }
-    if model.lowercased().hasPrefix("local/") { return true }
-    if model.lowercased() == "on-device" { return true }
-    return false
-  }
-
-  /// Pre-warms the TCP/TLS connection to OpenRouter by making a lightweight HEAD request.
-  /// This establishes and caches the connection for subsequent API calls.
-  /// Failures are logged but do not throw - warming is best-effort.
-  func warmUp() async {
-    let startTime = CFAbsoluteTimeGetCurrent()
-    let url = baseURL.appendingPathComponent("models")
-    var request = URLRequest(url: url)
-    request.httpMethod = "HEAD"
-    applyBrandHeaders(&request)
-
-    do {
-      let (_, response) = try await session.data(for: request)
-      let elapsed = (CFAbsoluteTimeGetCurrent() - startTime) * 1000
-      if let http = response as? HTTPURLResponse {
-        logger.info("Connection warm-up completed in \(elapsed, format: .fixed(precision: 1))ms (status: \(http.statusCode))")
-      } else {
-        logger.info("Connection warm-up completed in \(elapsed, format: .fixed(precision: 1))ms")
-      }
-    } catch {
-      let elapsed = (CFAbsoluteTimeGetCurrent() - startTime) * 1000
-      logger.warning("Connection warm-up failed after \(elapsed, format: .fixed(precision: 1))ms: \(error.localizedDescription, privacy: .public)")
-    }
-  }
-}
-
-private struct OpenRouterChatRequest: Encodable {
-  struct Message: Encodable {
-    let role: String
-    let content: String
-  }
-
-  let model: String
-  let temperature: Double
-  let messages: [Message]
-}
-
-private struct OpenRouterChatResponse: Decodable {
-  struct Choice: Decodable {
-    struct ChoiceMessage: Decodable {
-      let role: String
-      let content: String
-    }
-
-    let index: Int
-    let finish_reason: String?
-    let message: ChoiceMessage?
-  }
-
-  struct Usage: Decodable {
-    let prompt_tokens: Int
-    let completion_tokens: Int
-  }
-
-  let choices: [Choice]
-  let usage: Usage?
-}
-
-private struct OpenRouterValidationResponse: Decodable {
-  struct ValidationData: Decodable {
-    let valid: Bool?
-  }
-
-  let valid: Bool?
-  let data: ValidationData?
-}
-
-private struct OpenRouterStreamingChatRequest: Encodable {
-  struct Message: Encodable {
-    let role: String
-    let content: String
-  }
-
-  let model: String
-  let temperature: Double
-  let messages: [Message]
-  let stream: Bool
-}
-
-private struct OpenRouterAudioTranscriptionRequest: Encodable {
-  struct Message: Encodable {
-    let role: String
-    let content: [OpenRouterAudioContentPart]
-  }
-
-  let model: String
-  let temperature: Double
-  let messages: [Message]
-  let stream: Bool
-}
-
-private enum OpenRouterAudioContentPart: Encodable {
-  case text(String)
-  case inputAudio(data: String, format: String)
-
-  private enum CodingKeys: String, CodingKey {
-    case type
-    case text
-    case inputAudio = "input_audio"
-  }
-
-  func encode(to encoder: Encoder) throws {
-    var container = encoder.container(keyedBy: CodingKeys.self)
-    switch self {
-    case .text(let text):
-      try container.encode("text", forKey: .type)
-      try container.encode(text, forKey: .text)
-    case .inputAudio(let data, let format):
-      try container.encode("input_audio", forKey: .type)
-      let inputAudio = OpenRouterInputAudio(data: data, format: format)
-      try container.encode(inputAudio, forKey: .inputAudio)
-    }
-  }
-}
-
-private struct OpenRouterInputAudio: Encodable {
-  let data: String
-  let format: String
-}
-
-private struct OpenRouterStreamChunk: Decodable {
-  struct Choice: Decodable {
-    struct Delta: Decodable {
-      let role: String?
-      let content: String?
-    }
-    let delta: Delta?
-    let finish_reason: String?
-  }
-  let choices: [Choice]
-}
-
-// swiftlint:disable:this file_length
